Fixed an signed/unsigned warning.
git-svn-id: svn://svn.icculus.org/twilight/trunk/darkplaces@5393 d7cf8633-e32d-0410-b094-e92efae38249

+++ netconn.c
@@ -1234,7 +1234,7 @@ void NetConn_QueryQueueFrame(void)
 			continue;
 		}
 
-		if( entry->querycounter != net_slist_maxtries.integer ) 
+		if( entry->querycounter != (unsigned) net_slist_maxtries.integer ) 
 		{
 			lhnetaddress_t address;
 			int socket;
